Update 'First-order Logic'
This change removes the need for special eigenvalue constants. In fact,
they are not necessary as we can have the same without them if we
modify the definitions of satisfiability, entailment and validity of a
sequent and change the inference rules, so that a eigenvariable is
just an ordinary variable (of course with the same eigenvariable
condition). As of now, the introduction of these special eigenvalue
constants is not done properly, for example: how a sentence with such
an constant were to be evaluated if it contains a constant not in the
language? As it is in fact very unnatural to extend the language, the
proposed approach seems to be prefered.

Also this change fixes some proof trees and other small things, and
includes rewrites of some proofs, most notably the soundness proof.
The proofs should be revised, though, as the phrasing is clumsy and the
case analyses should be made clearer.

The change only includes the sequent calculus; the remaining chapters
would need to be adjusted as well.